fix: api 네이밍 수정과 로직 간편화
